+++ docs/CHANGELOG.md
@@ -35,6 +35,7 @@
 - `[Searchfield]` Fix on uneven searchfield in firefox. ([#5620](https://github.com/infor-design/enterprise/issues/5620))
 - `[Searchfield]` Fix on uneven searchfield in firefox. ([#5695](https://github.com/infor-design/enterprise/issues/5695))
 - `[Searchfield]` Fix on misaligned close button on mobile view. ([#5782](https://github.com/infor-design/enterprise/issues/5782))
+- `[Searchfield]` Change width when parent container becomes smaller. ([#4696](https://github.com/infor-design/enterprise/issues/4696))
 - `[Spinbox]` Remove functionality of Home and End buttons on Spinbox. ([#5659](https://github.com/infor-design/enterprise/issues/5659))
 - `[Spinbox]` Fix spinbox misalignment on sample sizes. ([#5733](https://github.com/infor-design/enterprise/issues/5733))
 - `[Tabs]` Fix a bug on vertical tabs scroll on panel containers. ([#5565](https://github.com/infor-design/enterprise/issues/5565))